requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic[email]
openai>=1.0.0
pillow>=10.0.0  
python-multipart>=0.0.6
aiofiles>=23.0.0
boto3
rembg
google-generativeai

aiofiles
aiohttp==3.9.1
httpx==0.25.2
dnspython==2.4.2
python-whois==0.8.0
aioredis==2.0.1
cryptography==42.0.5
# Domain Service Dependencies
razorpay==1.3.0
dnspython==2.4.2
requests==2.31.0

# Indian Payment Gateways

# Background Processing (Optional)
celery[redis]==5.4.0
redis==5.0.1

# Additional utilities
python-dotenv==1.0.0
pydantic-settings==2.0.3